# packages/contracts/openapi/argus-v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Argus Control Plane API",
    "version": "0.2.0",
    "description": "Execution pipeline and incident/status page contract. Public URL v1; ProbeJob and ProbeResult v0.2."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/monitors": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OrganizationHeader"
        }
      ],
      "get": {
        "operationId": "listMonitors",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of monitors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createMonitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonitorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Monitor created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/monitors/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OrganizationHeader"
        },
        {
          "$ref": "#/components/parameters/MonitorId"
        }
      ],
      "get": {
        "operationId": "getMonitor",
        "responses": {
          "200": {
            "description": "Monitor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "operationId": "updateMonitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonitorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated monitor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteMonitor",
        "responses": {
          "204": {
            "description": "Monitor deleted"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/monitors/{id}/timeseries": {
      "parameters": [
        {
          "$ref": "#/components/parameters/OrganizationHeader"
        },
        {
          "$ref": "#/components/parameters/MonitorId"
        }
      ],
      "get": {
        "operationId": "getMonitorTimeseries",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "1h",
                "6h",
                "24h"
              ],
              "default": "24h"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monitor time-series data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorTimeSeriesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "probeBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "argp_<probeId>.<secret>; HMAC digest only in storage"
      }
    },
    "parameters": {
      "OrganizationHeader": {
        "name": "X-Argus-Organization-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "MonitorId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Structured error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "HttpMonitorConfig": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind",
          "url",
          "method",
          "timeoutMs",
          "expectedStatus",
          "maxRedirects",
          "maxResponseBytes"
        ],
        "properties": {
          "kind": {
            "const": "http"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "pattern": "^https?://"
          },
          "method": {
            "enum": [
              "GET",
              "HEAD",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ]
          },
          "timeoutMs": {
            "type": "integer",
            "minimum": 100,
            "maximum": 30000
          },
          "expectedStatus": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "maxRedirects": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10
          },
          "maxResponseBytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1048576
          }
        }
      },
      "Monitor": {
        "type": "object",
        "required": [
          "id",
          "organizationId",
          "name",
          "intervalSeconds",
          "regions",
          "healthState",
          "version",
          "config",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy",
          "incidentPolicy"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "intervalSeconds": {
            "type": "integer",
            "minimum": 60
          },
          "regions": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string"
            }
          },
          "healthState": {
            "enum": [
              "UNKNOWN",
              "HEALTHY",
              "DEGRADED",
              "PENDING_DOWN",
              "DOWN",
              "PENDING_RECOVERY"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "config": {
            "$ref": "#/components/schemas/HttpMonitorConfig"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "incidentPolicy": {
            "$ref": "#/components/schemas/IncidentPolicy"
          }
        }
      },
      "CreateMonitorRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "intervalSeconds",
          "regions",
          "config"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "intervalSeconds": {
            "type": "integer",
            "minimum": 60
          },
          "regions": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string"
            }
          },
          "config": {
            "$ref": "#/components/schemas/HttpMonitorConfig"
          },
          "incidentPolicy": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "failureThreshold": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              },
              "recoveryThreshold": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              }
            }
          }
        }
      },
      "UpdateMonitorRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "intervalSeconds",
          "regions",
          "config",
          "version"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "intervalSeconds": {
            "type": "integer",
            "minimum": 60
          },
          "regions": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string"
            }
          },
          "config": {
            "$ref": "#/components/schemas/HttpMonitorConfig"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "incidentPolicy": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "failureThreshold": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              },
              "recoveryThreshold": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              }
            }
          }
        }
      },
      "MonitorPage": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Monitor"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {},
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "IncidentPolicy": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "failureThreshold",
          "recoveryThreshold"
        ],
        "properties": {
          "failureThreshold": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "recoveryThreshold": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          }
        }
      },
...
      "TimeSeriesPoint": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "executionId",
          "time",
          "kind",
          "outcome",
          "latencyMs"
        ],
        "properties": {
          "executionId": {
            "type": "string",
            "format": "uuid"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "kind": {
            "enum": [
              "SCHEDULED",
              "MANUAL",
              "EVALUATION"
            ]
          },
          "outcome": {
            "enum": [
              "PASS",
              "FAIL",
              null
            ]
          },
          "latencyMs": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "MonitorTimeSeriesResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "monitorId",
          "from",
          "to",
          "truncated",
          "series"
        ],
        "properties": {
          "monitorId": {
            "type": "string",
            "format": "uuid"
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "truncated": {
            "type": "boolean"
          },
          "series": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeSeriesPoint"
              }
            }
          }
        }
      }
    }
  }
}